+++ lib/email/architecture.rst
